+++ crates/client/src/admin_websocket.rs
@@ -8,6 +8,7 @@ use holochain_conductor_api::{
     PeerMetaInfo, SourceChainCursor, StorageInfo,
 };
 use holochain_types::network::HolochainTransportStats;
+use holochain_types::prelude::CapGrant;
 use holochain_types::websocket::AllowedOrigins;
 use holochain_types::{
     dna::AgentPubKey,
@@ -18,8 +19,8 @@ use holochain_types::{
 };
 use holochain_websocket::{connect, ConnectRequest, WebsocketConfig, WebsocketSender};
 use holochain_zome_types::prelude::{
-    CapAccess, DnaDef, GrantZomeCallCapabilityPayload, GrantedFunctions, ZomeCallCapGrant,
-    CAP_SECRET_BYTES,
+    Capability, DnaDef, GrantConstraint, GrantZomeCallCapabilityPayload, GrantedFunctions,
+    ZomeCallGrant, CAP_SECRET_BYTES,
 };
 use kitsune2_api::Url;
 use serde::{Deserialize, Serialize};
@@ -620,13 +621,15 @@ impl AdminWebsocket {
 
         self.grant_zome_call_capability(GrantZomeCallCapabilityPayload {
             cell_id: request.cell_id,
-            cap_grant: ZomeCallCapGrant {
+            cap_grant: CapGrant {
                 tag: "zome-call-signing-key".to_string(),
-                access: CapAccess::Assigned {
+                constraint: GrantConstraint::Assigned {
                     secret: cap_secret.into(),
                     assignees: BTreeSet::from([signing_agent_key.clone()]),
                 },
-                functions: request.functions.unwrap_or(GrantedFunctions::All),
+                capability: Capability::ZomeCall(ZomeCallGrant {
+                    functions: request.functions.unwrap_or(GrantedFunctions::All),
+                }),
             },
         })
         .await?;