+++ crates/hc_client/src/zome_call.rs
@@ -79,9 +79,8 @@ use holochain_client::{
 };
 use holochain_conductor_api::{CellInfo, IssueAppAuthenticationTokenPayload};
 use holochain_types::prelude::{
-    AgentPubKey, CapAccess, CapSecret, DnaHashB64, ExternIO, FunctionName,
-    GrantZomeCallCapabilityPayload, GrantedFunctions, InstalledAppId, ZomeCallCapGrant, ZomeName,
-    CAP_SECRET_BYTES,
+    AgentPubKey, CapGrant, CapSecret, DnaHashB64, ExternIO, FunctionName, GrantConstraint,
+    GrantZomeCallCapabilityPayload, GrantedFunctions, InstalledAppId, ZomeName, CAP_SECRET_BYTES,
 };
 use holochain_types::websocket::AllowedOrigins;
 use serde::{Deserialize, Serialize};
@@ -180,9 +179,9 @@ pub async fn zome_call_auth(zome_call_auth: ZomeCallAuth) -> anyhow::Result<()> 
         admin_client
             .grant_zome_call_capability(GrantZomeCallCapabilityPayload {
                 cell_id: cell_id.clone(),
-                cap_grant: ZomeCallCapGrant::new(
+                cap_grant: CapGrant::new_zome_call_grant(
                     "sandbox".to_string(),
-                    CapAccess::Assigned {
+                    GrantConstraint::Assigned {
                         secret: auth.cap_secret,
                         assignees: vec![signing_agent_key.clone()].into_iter().collect(),
                     },