+++ crates/holochain/src/core/workflow/initialize_zomes_workflow.rs
@@ -307,10 +307,12 @@ mod tests {
                 EntryVisibility::Private,
                 Entry::CapGrant(CapGrantEntry {
                     tag: "".into(),
-                    access: ().into(),
-                    functions: GrantedFunctions::Listed(
-                        vec![("no-init".into(), "xxx".into())].into_iter().collect(),
-                    ),
+                    constraint: ().into(),
+                    capability: Capability::ZomeCall(ZomeCallGrant {
+                        functions: GrantedFunctions::Listed(
+                            vec![("no-init".into(), "xxx".into())].into_iter().collect(),
+                        ),
+                    }),
                 }),
                 ChainTopOrdering::default(),
             ))?;