+++ crates/test_utils/wasm/wasm_workspace/create_entry/src/coordinator.rs
@@ -111,9 +111,11 @@ fn init(_: ()) -> ExternResult<InitCallbackResult> {
     let functions = GrantedFunctions::Listed(fns);
     create_cap_grant(CapGrantEntry {
         tag: "".into(),
-        // empty access converts to unrestricted
-        access: ().into(),
-        functions,
+        // empty constraint converts to unrestricted
+        constraint: ().into(),
+        capability: Capability::ZomeCall(ZomeCallGrant {
+            functions,
+        })
     })?;
 
     Ok(InitCallbackResult::Pass)