+++ crates/test_utils/wasm/wasm_workspace/foo/src/lib.rs
@@ -10,9 +10,11 @@ fn init() -> ExternResult<InitCallbackResult> {
     let functions = GrantedFunctions::Listed(fns);
     create_cap_grant(CapGrantEntry {
         tag: "".into(),
-        // empty access converts to unrestricted
-        access: ().into(),
-        functions,
+        // empty constraint converts to unrestricted
+        constraint: ().into(),
+        capability: Capability::ZomeCall(ZomeCallGrant {
+            functions,
+        })
     })?;
 
     Ok(InitCallbackResult::Pass)