+++ crates/test_utils/wasm/wasm_workspace/post_commit_volley/src/coordinator.rs
@@ -9,9 +9,11 @@ fn set_access(_: ()) -> ExternResult<()> {
     let functions = GrantedFunctions::Listed(fns);
     create_cap_grant(CapGrantEntry {
         tag: "".into(),
-        // empty access converts to unrestricted
-        access: ().into(),
-        functions,
+        // empty constraint converts to unrestricted
+        constraint: ().into(),
+        capability: Capability::ZomeCall(ZomeCallGrant {
+            functions,
+        })
     })?;
 
     Ok(())